+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.0 (June 07, 2020)
+  - Merge branch 'feature/readme' into develop
+  - completed documentation
+  - update checklist
+  - Merge branch 'feature/api' into develop
+  - Added User API: Delete user by username
+  - added User API: Get user by username
+  - added sample to get data with API
+  - Merge branch 'release-0.2' into develop
+
 ## 0.2.0 (June 02, 2020)
   - Merge branch 'feature/base-code' into develop
   - updated requirements.txt file